% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_transform.R
\name{apply_config_filters}
\alias{apply_config_filters}
\title{Apply filters defined in dataset configuration}
\usage{
apply_config_filters(data, filters)
}
\arguments{
\item{data}{List of data frames}

\item{filters}{Named list of filter values}
}
\value{
List of filtered data frames
}
\description{
Apply filters defined in dataset configuration
}
\keyword{internal}
