% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-config.R
\name{apply_mutate}
\alias{apply_mutate}
\title{Apply column mutations from configuration}
\usage{
apply_mutate(df, mutate_config)
}
\arguments{
\item{df}{Tibble, specifying data to transform}

\item{mutate_config}{List, specifying column creation rules}
}
\value{
Tibble with new columns added
}
\description{
Creates new columns from existing data (like dplyr::mutate)
}
\keyword{internal}
