% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\name{calculate_data_hash}
\alias{calculate_data_hash}
\title{Calculate SHA256 hash of data}
\usage{
calculate_data_hash(data)
}
\arguments{
\item{data}{Data frame or tibble, containing data to hash}
}
\value{
Character hash
}
\description{
Calculate SHA256 hash of data
}
\keyword{internal}
