% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_cache.R
\name{check_dataset_versions}
\alias{check_dataset_versions}
\title{Check for dataset version updates}
\usage{
check_dataset_versions()
}
\value{
Character vector of dataset names with updates available
}
\description{
Check for dataset version updates
}
\keyword{internal}
