% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-utils.R
\name{clean_column_values}
\alias{clean_column_values}
\title{Clean column values to snake_case}
\usage{
clean_column_values(df, column_names = NULL)
}
\arguments{
\item{df}{Tibble, specifying data with columns to clean}

\item{column_names}{Character vector, specifying column names to clean (e.g., c("measure", "statistic"))}
}
\value{
Tibble with cleaned column values
}
\description{
Applies clean_str to values in specified columns
}
\keyword{internal}
