% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-utils.R
\name{clean_org_names}
\alias{clean_org_names}
\title{Clean NHS Talking Therapies provider names}
\usage{
clean_org_names(x)
}
\arguments{
\item{x}{Character, specifying NHS TT provider name in all caps}
}
\value{
Clean NHS TT provider name
}
\description{
The provider names are all caps in the data, for visualisations they should be clean.
}
\keyword{internal}
