% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-utils.R
\name{clean_str}
\alias{clean_str}
\title{Clean strings to snake_case}
\usage{
clean_str(x)
}
\arguments{
\item{x}{Character, specifying string to clean (e.g., "CamelCase" or "Column Name")}
}
\value{
Clean string
}
\description{
Clean strings to snake_case
}
\keyword{internal}
