% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-utils.R
\name{convert_to_numeric}
\alias{convert_to_numeric}
\title{Convert measure columns to numeric}
\usage{
convert_to_numeric(df, measure_cols)
}
\arguments{
\item{df}{Tibble, specifying data to tidy}

\item{measure_cols}{Character vector, specifying measure column names}
}
\value{
Tibble with measure columns as numeric
}
\description{
Converts NULL to NA and coerces measure columns to numeric
}
\keyword{internal}
