% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-sources.R
\name{derive_tidy_source_url}
\alias{derive_tidy_source_url}
\title{Derive the GitHub Release URL for a tidy source}
\usage{
derive_tidy_source_url(dataset, version)
}
\arguments{
\item{dataset}{Character, dataset name}

\item{version}{Character, dataset version}
}
\value{
Character URL
}
\description{
Derive the GitHub Release URL for a tidy source
}
\keyword{internal}
