% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Package attach hook}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Library name}

\item{pkgname}{Package name}
}
\description{
Silently sets up package on first load, only shows message if initialization fails
}
\keyword{internal}
