% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\name{download_and_store}
\alias{download_and_store}
\title{Download, parse, and store raw data}
\usage{
download_and_store(
  dataset,
  period,
  frequency,
  url,
  source_format,
  file_path,
  csv_pattern = NULL,
  sheet = NULL,
  range = NULL
)
}
\arguments{
\item{dataset}{Character, dataset name}

\item{period}{Character, reporting period}

\item{frequency}{Character, "annual", "quarterly" or "monthly"}

\item{url}{Character, source URL}

\item{source_format}{Character, source format ("csv", "zip", "rar", "xlsx")}

\item{file_path}{Character, destination path}

\item{csv_pattern}{Character, regex to locate CSV inside archive (optional)}

\item{sheet}{Character, sheet name or index for Excel sources (required for Excel sources)}

\item{range}{Character, Excel cell range (e.g., "A5:H427") for Excel sources (optional)}
}
\value{
Character path to cached file
}
\description{
Download, parse, and store raw data
}
\keyword{internal}
