% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\name{download_raw}
\alias{download_raw}
\title{Download and cache raw data}
\usage{
download_raw(dataset, period, frequency, use_cache = TRUE)
}
\arguments{
\item{dataset}{Character, specifying dataset name (e.g., "measures_annual", "measures_monthly")}

\item{period}{Character, specifying reporting period (e.g., "2023-24" for annual, "2026-27-q1" for quarterly, "2025-09" for monthly)}

\item{frequency}{Character, specifying report frequency ("annual", "quarterly" or "monthly")}

\item{use_cache}{Logical, specifying whether to use cached data if available. Default TRUE}
}
\value{
Character path to cached file
}
\description{
Download and cache raw data
}
\keyword{internal}
