% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\name{download_with_retry}
\alias{download_with_retry}
\title{Download data from URL with retry logic}
\usage{
download_with_retry(url, file)
}
\arguments{
\item{url}{Character, specifying URL to download from}

\item{file}{Character, specifying destination file path}
}
\description{
Download data from URL with retry logic
}
\keyword{internal}
