% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\name{extract_csv_from_archive}
\alias{extract_csv_from_archive}
\title{Extract CSV from archive using pattern matching}
\usage{
extract_csv_from_archive(archive_path, csv_pattern)
}
\arguments{
\item{archive_path}{Character, specifying path to zip/rar archive}

\item{csv_pattern}{Character, specifying regex pattern to match CSV filename}
}
\value{
Tibble with raw data
}
\description{
Extract CSV from archive using pattern matching
}
\keyword{internal}
