% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-utils.R
\name{filter_rows}
\alias{filter_rows}
\title{Filter rows}
\usage{
filter_rows(df, filter_config)
}
\arguments{
\item{df}{Tibble, specifying data to filter}

\item{filter_config}{Named list, specifying filter values (e.g., list(org_type = c("Provider")))}
}
\value{
Filtered tibble
}
\description{
Filters rows based on configuration
}
\keyword{internal}
