% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_cache.R
\name{format_date_range}
\alias{format_date_range}
\title{Format date range for display}
\usage{
format_date_range(start_date, end_date)
}
\arguments{
\item{start_date}{Date or character start date}

\item{end_date}{Date or character end date}
}
\value{
Character formatted date range
}
\description{
Format date range for display
}
\keyword{internal}
