% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_cache.R
\name{format_iso_datetime}
\alias{format_iso_datetime}
\title{Format POSIXct as ISO-8601 string}
\usage{
format_iso_datetime(x)
}
\arguments{
\item{x}{POSIXct time}
}
\value{
Character string in UTC
}
\description{
Format POSIXct as ISO-8601 string
}
\keyword{internal}
