% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{format_raw_cache}
\alias{format_raw_cache}
\title{Format raw cache summary for display}
\usage{
format_raw_cache(summary)
}
\arguments{
\item{summary}{List, from summarise_raw_cache()}
}
\value{
Character, e.g. "1.2 Mb (3 downloads)"
}
\description{
Format raw cache summary for display
}
\keyword{internal}
