% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_datasets.R
\name{get_annual_dataset_configs}
\alias{get_annual_dataset_configs}
\title{Get dataset configurations for annual reports}
\usage{
get_annual_dataset_configs()
}
\value{
List of dataset configurations
}
\description{
Configurations for datasets from annual report zip files.
Each annual zip contains multiple CSV files that are extracted using patterns.
}
\keyword{internal}
