% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_sources.R
\name{get_annual_report_links}
\alias{get_annual_report_links}
\title{Get data source links for annual reports}
\usage{
get_annual_report_links()
}
\value{
Named list of URLs by period
}
\description{
URLs for annual report zip files. Each zip contains multiple CSV files.
}
\keyword{internal}
