% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_sources.R
\name{get_available_periods}
\alias{get_available_periods}
\title{Get available periods for a source type}
\usage{
get_available_periods(source_type = "annual_reports")
}
\arguments{
\item{source_type}{Character, the data source type}
}
\value{
Character vector of available periods
}
\description{
Get available periods for a source type
}
\keyword{internal}
