% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhstt_data.R
\name{get_cache_dir}
\alias{get_cache_dir}
\title{Get package cache directory}
\usage{
get_cache_dir()
}
\value{
Character, the path to the cache directory
}
\description{
Get package cache directory
}
\keyword{internal}
