% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_cache.R
\name{get_cache_index}
\alias{get_cache_index}
\title{Retrieve cache index information}
\usage{
get_cache_index()
}
\value{
List containing cache metadata or NULL if missing
}
\description{
Retrieve cache index information
}
\keyword{internal}
