% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_cache.R
\name{get_cache_index_path}
\alias{get_cache_index_path}
\title{Cache index path}
\usage{
get_cache_index_path(cache_dir = get_cache_dir())
}
\arguments{
\item{cache_dir}{Cache directory}
}
\value{
Character path to cache index file
}
\description{
Cache index path
}
\keyword{internal}
