% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_datasets.R
\name{get_dataset_configs}
\alias{get_dataset_configs}
\title{Get all dataset configurations}
\usage{
get_dataset_configs()
}
\value{
List of all dataset configurations
}
\description{
Combines all dataset configurations from different sources.
}
\keyword{internal}
