% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-config.R
\name{get_dataset_version}
\alias{get_dataset_version}
\title{Get dataset version}
\usage{
get_dataset_version(dataset, frequency)
}
\arguments{
\item{dataset}{Character, specifying dataset name (e.g., "key_measures", "activity_performance")}

\item{frequency}{Character, specifying report frequency ("annual" or "monthly")}
}
\value{
Character version string
}
\description{
Get dataset version
}
\keyword{internal}
