% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-metadata-providers.R
\name{get_metadata_providers}
\alias{get_metadata_providers}
\title{Get provider organisation metadata from ODS}
\usage{
get_metadata_providers(use_cache = TRUE)
}
\arguments{
\item{use_cache}{Logical, whether to use cached data if available. Default TRUE.}
}
\value{
Tibble with provider organisation metadata
}
\description{
Get provider organisation metadata from ODS
}
\examples{
\dontrun{
# Get provider metadata
nhstt_providers <- get_metadata_providers()
}
}
