% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_sources.R
\name{get_package_version}
\alias{get_package_version}
\title{Get package version}
\usage{
get_package_version()
}
\value{
Character, package version string
}
\description{
Get package version
}
\keyword{internal}
