% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_sources.R
\name{get_source_links}
\alias{get_source_links}
\title{Get all source links by type}
\usage{
get_source_links(source_type)
}
\arguments{
\item{source_type}{Character, the type of source}
}
\value{
Named list of URLs
}
\description{
Get all source links by type
}
\keyword{internal}
