% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_tidy_cache_path}
\alias{get_tidy_cache_path}
\title{Get versioned tidy cache path}
\usage{
get_tidy_cache_path(dataset, period, frequency, dataset_version = NULL)
}
\arguments{
\item{dataset}{Character, specifying dataset name (e.g., "key_measures_annual", "activity_performance_monthly")}

\item{period}{Character, specifying reporting period (e.g., "2023-24" for annual, "2025-09" for monthly)}

\item{frequency}{Character, specifying report frequency ("annual" or "monthly")}

\item{dataset_version}{Character, specifying dataset version (e.g., "1.0.0"). Default NULL}
}
\value{
Character path to cached Parquet file
}
\description{
Get versioned tidy cache path
}
\keyword{internal}
