% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-sources.R
\name{get_tidy_source_sidecar_path}
\alias{get_tidy_source_sidecar_path}
\title{Get path to tidy source version sidecar JSON}
\usage{
get_tidy_source_sidecar_path(dataset, version)
}
\arguments{
\item{dataset}{Character, dataset name}

\item{version}{Character, dataset version}
}
\value{
Character path
}
\description{
Get path to tidy source version sidecar JSON
}
\keyword{internal}
