% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-sources.R
\name{invalidate_tidy_source_cache}
\alias{invalidate_tidy_source_cache}
\title{Remove a versioned tidy source cache file}
\usage{
invalidate_tidy_source_cache(dataset, version)
}
\arguments{
\item{dataset}{Character, dataset name}

\item{version}{Character, dataset version}
}
\value{
Invisible TRUE
}
\description{
Remove a versioned tidy source cache file
}
\keyword{internal}
