% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{is_roxygen_context}
\alias{is_roxygen_context}
\title{Detect if package is being loaded by roxygen2}
\usage{
is_roxygen_context()
}
\value{
Logical, TRUE if roxygen2 is in the call stack
}
\description{
Checks the call stack for roxygen2 or as.list.environment calls,
which indicates documentation is being generated.
}
\keyword{internal}
