% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhstt_data.R
\name{is_setup}
\alias{is_setup}
\title{Check if package data is set up}
\usage{
is_setup()
}
\value{
Logical, returns TRUE if NHS TT data is set up, FALSE otherwise
}
\description{
Check if package data is set up
}
\keyword{internal}
