% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-config.R
\name{list_available_periods}
\alias{list_available_periods}
\title{List available periods for a dataset and frequency}
\usage{
list_available_periods(dataset, frequency, include_development = FALSE)
}
\arguments{
\item{dataset}{Character, specifying dataset name}

\item{frequency}{Character, specifying report frequency ("annual", "quarterly" or "monthly")}

\item{include_development}{Logical, specifying whether to include periods marked with development = true. Default FALSE}
}
\value{
Character vector of available periods
}
\description{
List available periods for a dataset and frequency
}
\keyword{internal}
