% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-metadata.R
\name{load_combined_metadata}
\alias{load_combined_metadata}
\title{Helper to load and combine main and additional metadata sheets}
\usage{
load_combined_metadata(dataset_base, period, frequency, use_cache)
}
\arguments{
\item{dataset_base}{Base name ("metadata_measures" or "metadata_variables")}

\item{period}{Reporting period}

\item{frequency}{Frequency ("annual")}

\item{use_cache}{Whether to use cache}
}
\value{
Combined tibble
}
\description{
Helper to load and combine main and additional metadata sheets
}
\keyword{internal}
