% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{load_tidy_cache}
\alias{load_tidy_cache}
\title{Load tidy data from Parquet cache}
\usage{
load_tidy_cache(dataset, period, frequency)
}
\arguments{
\item{dataset}{Character, specifying dataset name (e.g., "key_measures", "activity_performance")}

\item{period}{Character, specifying reporting period (e.g., "2023-24" for annual, "2025-09" for monthly)}

\item{frequency}{Character, specifying report frequency ("annual" or "monthly")}
}
\value{
Tibble
}
\description{
Load tidy data from Parquet cache
}
\keyword{internal}
