% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-config.R
\name{load_tidy_config}
\alias{load_tidy_config}
\title{Load all tidy configurations}
\usage{
load_tidy_config()
}
\value{
List containing all tidy configurations
}
\description{
Load all tidy configurations
}
\keyword{internal}
