% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-utils.R
\name{make_clean_str}
\alias{make_clean_str}
\title{Make clean strings}
\usage{
make_clean_str(x)
}
\arguments{
\item{x}{Character, specifying string to clean (e.g., "CamelCase")}
}
\value{
Clean string
}
\description{
Make clean strings
}
\keyword{internal}
