% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-config.R
\name{mutate_columns}
\alias{mutate_columns}
\title{Create new columns from existing data}
\usage{
mutate_columns(df, mutate_config)
}
\arguments{
\item{df}{Tibble, specifying data to transform}

\item{mutate_config}{List, specifying column creation rules}
}
\value{
Tibble with new columns added
}
\description{
Creates new columns from existing data (like dplyr::mutate)
}
\keyword{internal}
