% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhstt_setup.R
\name{nhstt_clear_download_cache}
\alias{nhstt_clear_download_cache}
\title{Clear cached data}
\usage{
nhstt_clear_download_cache(confirm = TRUE)
}
\arguments{
\item{confirm}{Logical, specifying whether to require confirmation before deletion.
Default is TRUE.}
}
\value{
Invisible NULL
}
\description{
Clear cached data
}
\examples{
\donttest{
nhstt_clear_download_cache(confirm = FALSE)
}
}
