% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhstt_data.R
\name{nhstt_data}
\alias{nhstt_data}
\title{Get NHS TT data}
\usage{
nhstt_data(dataset = "key_measures")
}
\arguments{
\item{dataset}{Character, specifying which dataset to load.
Use \code{\link[=nhstt_info]{nhstt_info()}} to see options.
Default is "key_measures".}
}
\value{
A tibble containing the requested dataset
}
\description{
Load tidied NHS Talking Therapies data from cache. After running
\code{\link[=nhstt_setup]{nhstt_setup()}}, datasets can also be accessed directly by name
(e.g., \code{key_measures}) instead of using this function.
}
\keyword{internal}
