% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhstt_info.R
\name{nhstt_info}
\alias{nhstt_info}
\title{Get information about NHS Talking Therapies datasets}
\usage{
nhstt_info()
}
\value{
Invisible list containing metadata information
}
\description{
Shows all available datasets in the package, including their version, description,
download status, and whether updates are available.
}
\examples{
nhstt_info()
}
