% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_validate.R
\name{parse_annual_period_bounds}
\alias{parse_annual_period_bounds}
\title{Parse annual (financial year) period bounds}
\usage{
parse_annual_period_bounds(periods)
}
\arguments{
\item{periods}{Character vector like "fy2324"}
}
\value{
List with start_date and end_date
}
\description{
Parse annual (financial year) period bounds
}
\keyword{internal}
