% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_cache.R
\name{parse_iso_datetime}
\alias{parse_iso_datetime}
\title{Parse ISO-8601 string into POSIXct}
\usage{
parse_iso_datetime(x)
}
\arguments{
\item{x}{Character string}
}
\value{
POSIXct time in UTC
}
\description{
Parse ISO-8601 string into POSIXct
}
\keyword{internal}
