% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-utils.R
\name{parse_monthly_period_bounds}
\alias{parse_monthly_period_bounds}
\title{Parse monthly period bounds}
\usage{
parse_monthly_period_bounds(period)
}
\arguments{
\item{period}{Character, specifying monthly period (e.g., "2025-09")}
}
\value{
List with start and end dates
}
\description{
Converts ISO year-month format to start/end dates of that month
}
\keyword{internal}
