% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-utils.R
\name{recode_values}
\alias{recode_values}
\title{Recode values}
\usage{
recode_values(df, recode_config)
}
\arguments{
\item{df}{Tibble, specifying data with columns to recode}

\item{recode_config}{Named list, specifying for each column a mapping of
\code{new_value: old_value} (e.g., list(variable_a = list(None = "null")))}
}
\value{
Tibble with recoded values
}
\description{
Replaces values in columns using a mapping (e.g., for a category that is
labelled differently in some periods)
}
\keyword{internal}
