% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-utils.R
\name{replace_suppression_with_na}
\alias{replace_suppression_with_na}
\title{Replace NHS suppression markers with NA}
\usage{
replace_suppression_with_na(x)
}
\arguments{
\item{x}{Character or numeric vector, specifying values to convert}
}
\value{
Numeric vector with suppression markers as NA
}
\description{
Converts NHS suppression markers ("*", "-", "NULL", etc.) to NA and coerces to numeric
}
\keyword{internal}
