% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-utils.R
\name{replace_values_with_na}
\alias{replace_values_with_na}
\title{Replace values with NA}
\usage{
replace_values_with_na(df, values)
}
\arguments{
\item{df}{Tibble, specifying data to clean}

\item{values}{Character vector, specifying values to replace with NA}
}
\value{
Tibble with placeholder values as NA
}
\description{
Converts placeholder values (e.g., "null", "All_ICB") to NA in all
character columns
}
\keyword{internal}
