% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\name{resolve_cache_file}
\alias{resolve_cache_file}
\title{Check if cache file exists}
\usage{
resolve_cache_file(file_path, use_cache)
}
\arguments{
\item{file_path}{Character, path to cache file}

\item{use_cache}{Logical, whether to use cached data if present}
}
\value{
Character path if cache exists and should be used, NULL otherwise
}
\description{
Check if cache file exists
}
\keyword{internal}
