% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-config.R
\name{resolve_periods}
\alias{resolve_periods}
\title{Resolve periods argument}
\usage{
resolve_periods(periods, dataset, frequency)
}
\arguments{
\item{periods}{Character vector or NULL, specifying periods (e.g., c("2023-24", "2024-25") for annual, c("2025-08", "2025-09") for monthly). Default NULL returns all periods}

\item{dataset}{Character, specifying dataset name (e.g., "key_measures_annual", "activity_performance_monthly")}

\item{frequency}{Character, specifying report frequency ("annual" or "monthly")}
}
\value{
Character vector of validated periods
}
\description{
Resolve periods argument
}
\keyword{internal}
