% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_select.R
\name{select_tables_by_pattern}
\alias{select_tables_by_pattern}
\title{Select tables by pattern}
\usage{
select_tables_by_pattern(raw_data, pattern)
}
\arguments{
\item{raw_data}{Named list of data frames by period}

\item{pattern}{Regex pattern to match table names}
}
\value{
Filtered list
}
\description{
Select tables by pattern
}
\keyword{internal}
