% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_transform.R
\name{standardise_variables}
\alias{standardise_variables}
\title{Standardise variable names}
\usage{
standardise_variables(data, var_mapping)
}
\arguments{
\item{data}{Nested list of data frames}

\item{var_mapping}{Named vector mapping old names to new names}
}
\value{
List with renamed variables
}
\description{
Standardise variable names
}
\keyword{internal}
