% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-utils.R
\name{str_extract_icd10}
\alias{str_extract_icd10}
\title{Extract ICD-10 codes from text}
\usage{
str_extract_icd10(x)
}
\arguments{
\item{x}{A character vector containing text}
}
\value{
A character vector with extracted standardised ICD-10 codes (comma-separated with no decimal point) or NA
}
\description{
To extract ICD-10 codes we use a regex pattern that matches the standard ICD-10 format:
a letter (A-Z) followed by two digits, optionally followed by a decimal point and 1-2 more digits.
}
\keyword{internal}
