% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-utils.R
\name{str_extract_snomed}
\alias{str_extract_snomed}
\title{Extract SNOMED CT codes from text}
\usage{
str_extract_snomed(x)
}
\arguments{
\item{x}{A character vector containing text}
}
\value{
A list of character vectors
}
\description{
To extract SNOMED CT codes we use a regex pattern that matches 6 to 18 digit numbers that don't start with zero.
}
\keyword{internal}
