% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{summarise_raw_cache}
\alias{summarise_raw_cache}
\title{Summarise raw cache for a frequency}
\usage{
summarise_raw_cache(frequency)
}
\arguments{
\item{frequency}{Character, specifying report frequency ("annual", "quarterly" or "monthly")}
}
\value{
List with size (bytes), count (downloads) and downloads (metadata)
}
\description{
Summarise raw cache for a frequency
}
\keyword{internal}
