% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{theme_nhstt}
\alias{theme_nhstt}
\title{ggplot2 theme for NHS TT figures}
\usage{
theme_nhstt(base_size = 13)
}
\arguments{
\item{base_size}{Numeric, base font size. Default 13}
}
\value{
A ggplot2 theme object
}
\description{
Applied internally by \code{\link[=plot_measures]{plot_measures()}} and \code{\link[=plot_measures_deciles]{plot_measures_deciles()}}.
}
\keyword{internal}
