% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_annual.R
\name{tidy_annual_dataset}
\alias{tidy_annual_dataset}
\title{Tidy annual dataset}
\usage{
tidy_annual_dataset(raw_data, config)
}
\arguments{
\item{raw_data}{Named list of raw data by period}

\item{config}{Dataset configuration list}
}
\value{
Tidied tibble
}
\description{
Generic tidying function for annual report datasets. Applies the standard
tidying pipeline using configuration provided in the dataset config.
}
\keyword{internal}
