% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-monthly.R
\name{tidy_config_activity_performance}
\alias{tidy_config_activity_performance}
\title{Tidy configuration for activity performance dataset}
\usage{
tidy_config_activity_performance()
}
\value{
List containing tidy configuration for activity_performance dataset
}
\description{
Tidy configuration for activity performance dataset
}
\keyword{internal}
