% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-annual.R
\name{tidy_config_key_measures}
\alias{tidy_config_key_measures}
\title{Tidy configuration for key measures dataset}
\usage{
tidy_config_key_measures()
}
\value{
List containing tidy configuration for key_measures dataset
}
\description{
Tidy configuration for key measures dataset
}
\keyword{internal}
