% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-metadata.R
\name{tidy_config_metadata}
\alias{tidy_config_metadata}
\title{Tidy configuration for metadata dataset}
\usage{
tidy_config_metadata()
}
\value{
List containing tidy configuration for metadata dataset
}
\description{
Tidy configuration for metadata dataset
}
\keyword{internal}
