% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-metadata.R
\name{tidy_config_metadata_measures_additional}
\alias{tidy_config_metadata_measures_additional}
\title{Tidy configuration for annual metadata measures (additional)}
\usage{
tidy_config_metadata_measures_additional()
}
\value{
List containing tidy configuration for metadata_measures_additional dataset
}
\description{
Tidy configuration for annual metadata measures (additional)
}
\keyword{internal}
