% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-metadata.R
\name{tidy_config_metadata_measures_main}
\alias{tidy_config_metadata_measures_main}
\title{Tidy configuration for annual metadata measures (main)}
\usage{
tidy_config_metadata_measures_main()
}
\value{
List containing tidy configuration for metadata_measures_main dataset
}
\description{
Tidy configuration for annual metadata measures (main)
}
\keyword{internal}
