% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-metadata.R
\name{tidy_config_metadata_variables_additional}
\alias{tidy_config_metadata_variables_additional}
\title{Tidy configuration for annual metadata variables (additional)}
\usage{
tidy_config_metadata_variables_additional()
}
\value{
List containing tidy configuration for metadata_variables_additional dataset
}
\description{
Tidy configuration for annual metadata variables (additional)
}
\keyword{internal}
