% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-metadata.R
\name{tidy_config_metadata_variables_main}
\alias{tidy_config_metadata_variables_main}
\title{Tidy configuration for annual metadata variables (main)}
\usage{
tidy_config_metadata_variables_main()
}
\value{
List containing tidy configuration for metadata_variables_main dataset
}
\description{
Tidy configuration for annual metadata variables (main)
}
\keyword{internal}
