% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-config.R
\name{tidy_config_registry}
\alias{tidy_config_registry}
\title{Registry of tidy configuration functions}
\usage{
tidy_config_registry()
}
\value{
Named list of tidy configuration functions
}
\description{
Registry of tidy configuration functions
}
\keyword{internal}
