% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-utils.R
\name{tidy_numeric_values}
\alias{tidy_numeric_values}
\title{Tidy numeric values}
\usage{
tidy_numeric_values(x)
}
\arguments{
\item{x}{Character or numeric vector, specifying values to tidy}
}
\value{
Numeric vector with suppression markers as NA
}
\description{
Converts NHS suppression markers to NA
}
\keyword{internal}
