% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_cache.R
\name{update_cache_index}
\alias{update_cache_index}
\title{Update cache index after setup}
\usage{
update_cache_index(cache_dir, configs, datasets_updated)
}
\arguments{
\item{cache_dir}{Cache directory path}

\item{configs}{Dataset configuration list}

\item{datasets_updated}{Character vector of dataset names updated}
}
\value{
Invisible NULL
}
\description{
Update cache index after setup
}
\keyword{internal}
