% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-config.R
\name{validate_archives}
\alias{validate_archives}
\title{Validate archives section}
\usage{
validate_archives(archives)
}
\arguments{
\item{archives}{List, specifying archives configuration}
}
\value{
Invisible TRUE if valid, aborts otherwise
}
\description{
Validate archives section
}
\keyword{internal}
