% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-config.R
\name{validate_dataset_config}
\alias{validate_dataset_config}
\title{Validate individual dataset configuration}
\usage{
validate_dataset_config(dataset_name, dataset, config)
}
\arguments{
\item{dataset_name}{Character, specifying dataset name}

\item{dataset}{List, specifying dataset configuration}

\item{config}{List, specifying full configuration (for archive lookups)}
}
\value{
Invisible TRUE if valid, aborts otherwise
}
\description{
Validate individual dataset configuration
}
\keyword{internal}
