% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-config.R
\name{validate_frequency}
\alias{validate_frequency}
\title{Validate frequency}
\usage{
validate_frequency(frequency)
}
\arguments{
\item{frequency}{Character, specifying report frequency to validate ("annual", "quarterly", "monthly", or "live")}
}
\value{
Invisible TRUE if valid, aborts otherwise
}
\description{
Validate frequency
}
\keyword{internal}
