% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{validate_measures}
\alias{validate_measures}
\title{Validate a measures table}
\usage{
validate_measures(measures)
}
\arguments{
\item{measures}{Tibble, as returned by \code{\link[=create_measures]{create_measures()}}}
}
\value{
Invisible TRUE, aborts if required columns are missing
}
\description{
Validate a measures table
}
\keyword{internal}
