% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-config.R
\name{validate_period}
\alias{validate_period}
\title{Validate period for a dataset and frequency}
\usage{
validate_period(period, dataset, frequency)
}
\arguments{
\item{period}{Character, specifying reporting period to validate (e.g., "2023-24" for annual, "2026-27-q1" for quarterly, "2025-09" for monthly)}

\item{dataset}{Character, specifying dataset name (e.g., "measures_annual", "measures_monthly")}

\item{frequency}{Character, specifying report frequency ("annual", "quarterly" or "monthly")}
}
\value{
Invisible TRUE if valid, aborts otherwise
}
\description{
Validate period for a dataset and frequency
}
\keyword{internal}
