% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-config.R
\name{validate_raw_config}
\alias{validate_raw_config}
\title{Validate raw configuration structure}
\usage{
validate_raw_config(config)
}
\arguments{
\item{config}{List, specifying parsed raw configuration}
}
\value{
Invisible TRUE if valid, aborts otherwise
}
\description{
Validate raw configuration structure
}
\keyword{internal}
