% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-config.R
\name{validate_tidy_config}
\alias{validate_tidy_config}
\title{Validate tidy configuration structure}
\usage{
validate_tidy_config(config)
}
\arguments{
\item{config}{List, specifying parsed tidy configuration}
}
\value{
Invisible TRUE if valid, aborts otherwise
}
\description{
Validate tidy configuration structure
}
\keyword{internal}
