% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_cache.R
\name{write_cache_index}
\alias{write_cache_index}
\title{Write cache index to disk}
\usage{
write_cache_index(index, cache_dir = get_cache_dir())
}
\arguments{
\item{index}{Cache index list}

\item{cache_dir}{Cache directory}
}
\description{
Write cache index to disk
}
\keyword{internal}
