% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{write_raw_downloads_json}
\alias{write_raw_downloads_json}
\title{Record raw download metadata}
\usage{
write_raw_downloads_json(
  dataset,
  period,
  frequency,
  url,
  source_format,
  storage_format,
  raw_data_hash,
  file_size
)
}
\arguments{
\item{dataset}{Character, specifying dataset name (e.g., "measures_annual", "measures_monthly")}

\item{period}{Character, specifying reporting period (e.g., "2023-24" for annual, "2026-27-q1" for quarterly, "2025-09" for monthly)}

\item{frequency}{Character, specifying report frequency ("annual", "quarterly" or "monthly")}

\item{url}{Character, specifying source URL}

\item{source_format}{Character, specifying original format ("zip", "rar", "csv", "xlsx")}

\item{storage_format}{Character, specifying how it's stored ("csv", "parquet")}

\item{raw_data_hash}{Character, specifying SHA256 hash of data}

\item{file_size}{Numeric, specifying size of stored file in bytes}
}
\description{
Record raw download metadata
}
\keyword{internal}
