% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{write_tidy_cache}
\alias{write_tidy_cache}
\title{Write tidy data to Parquet cache}
\usage{
write_tidy_cache(data, dataset, period, frequency, raw_data_hash, raw_data_url)
}
\arguments{
\item{data}{Tibble, containing cleaned data}

\item{dataset}{Character, specifying dataset name (e.g., "key_measures_annual", "activity_performance_monthly")}

\item{period}{Character, specifying reporting period (e.g., "2023-24" for annual, "2025-09" for monthly)}

\item{frequency}{Character, specifying report frequency ("annual" or "monthly")}

\item{raw_data_hash}{Character, specifying SHA256 hash of raw data}

\item{raw_data_url}{Character, specifying URL of source data}
}
\description{
Write tidy data to Parquet cache
}
\keyword{internal}
