% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-config.R
\name{available_nhstt_reports}
\alias{available_nhstt_reports}
\title{List available NHS Talking Therapies reports}
\usage{
available_nhstt_reports()
}
\value{
Tibble with dataset and frequency information (one row per dataset-frequency combination)
}
\description{
Returns a tibble with information about available datasets including
their time period coverage and frequency
}
\examples{
available_nhstt_reports()
}
