% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-sources.R
\name{download_tidy_source}
\alias{download_tidy_source}
\title{Download pre-built tidy parquet and store in cache}
\usage{
download_tidy_source(dataset, url, version)
}
\arguments{
\item{dataset}{Character, dataset name}

\item{url}{Character, download URL}

\item{version}{Character, dataset version (stored in sidecar)}
}
\value{
Invisible path to cached parquet
}
\description{
Download pre-built tidy parquet and store in cache
}
\keyword{internal}
