% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{effect_size}
\alias{effect_size}
\title{Effect size dataset from NHS TT annual reports}
\format{
A tibble in tidy/long format with the following columns:
\describe{
\item{start_date}{TODO}
\item{end_date}{TODO}
\item{period}{TODO}
\item{org_type}{TODO}
\item{org_code}{TODO}
\item{org_name}{TODO}
\item{proms}{TODO}
\item{initial_caseness}{TODO}
\item{variable_type}{TODO}
\item{variable_a}{TODO}
\item{variable_b}{TODO}
\item{statistic}{TODO}
\item{measure}{TODO}
\item{value}{TODO}
}
}
\source{
\url{https://digital.nhs.uk/data-and-information/publications/statistical/nhs-talking-therapies-for-anxiety-and-depression-annual-reports}
}
\description{
TODO: Add detailed description of the effect size dataset.
}
\details{
After running \code{\link[=nhstt_setup]{nhstt_setup()}}, this dataset can be accessed directly by running
\code{effect_size}.
}
\examples{
\donttest{
effect_size
}
}
\keyword{datasets}
