% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-utils.R
\name{get_available_periods}
\alias{get_available_periods}
\title{Get available periods for an archive}
\usage{
get_available_periods(archive_sources)
}
\arguments{
\item{archive_sources}{List, archive sources configuration}
}
\value{
Character vector of periods sorted descending
}
\description{
Get available periods for an archive
}
\keyword{internal}
