% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_raw_cache_dir}
\alias{get_raw_cache_dir}
\title{Get raw data cache directory}
\usage{
get_raw_cache_dir(frequency)
}
\arguments{
\item{frequency}{Character, specifying report frequency ("annual" or "monthly")}
}
\value{
Character path to raw data directory
}
\description{
Get raw data cache directory
}
\keyword{internal}
