% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_raw_cache_path}
\alias{get_raw_cache_path}
\title{Get raw data cache path}
\usage{
get_raw_cache_path(dataset, period, frequency)
}
\arguments{
\item{dataset}{Character, specifying dataset name (e.g., "key_measures_annual", "activity_performance_monthly")}

\item{period}{Character, specifying reporting period (e.g., "2023-24" for annual, "2025-09" for monthly)}

\item{frequency}{Character, specifying report frequency ("annual" or "monthly")}
}
\value{
Character path to raw data file
}
\description{
Get raw data cache path
}
\keyword{internal}
