% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_raw_downloads_json_path}
\alias{get_raw_downloads_json_path}
\title{Get raw downloads metadata file path}
\usage{
get_raw_downloads_json_path(frequency)
}
\arguments{
\item{frequency}{Character, specifying report frequency ("annual" or "monthly")}
}
\value{
Character path to .downloads.json file
}
\description{
Returns the path to the JSON file that stores download metadata for raw data
}
\keyword{internal}
