% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-sources.R
\name{get_tidy_source_cache_path}
\alias{get_tidy_source_cache_path}
\title{Get path to pre-built tidy parquet in cache}
\usage{
get_tidy_source_cache_path(dataset)
}
\arguments{
\item{dataset}{Character, dataset name}
}
\value{
Character path
}
\description{
Get path to pre-built tidy parquet in cache
}
\keyword{internal}
