% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-config.R
\name{load_raw_config}
\alias{load_raw_config}
\title{Load raw data configuration from TOML}
\usage{
load_raw_config()
}
\value{
List containing parsed raw data configuration
}
\description{
Load raw data configuration from TOML
}
\keyword{internal}
