% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-sources.R
\name{load_tidy_source}
\alias{load_tidy_source}
\title{Load pre-built tidy parquet from cache}
\usage{
load_tidy_source(dataset)
}
\arguments{
\item{dataset}{Character, dataset name}
}
\value{
Tibble
}
\description{
Load pre-built tidy parquet from cache
}
\keyword{internal}
