% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-sources.R
\name{load_tidy_sources_config}
\alias{load_tidy_sources_config}
\title{Load tidy data sources configuration}
\usage{
load_tidy_sources_config()
}
\value{
List of dataset configurations keyed by dataset name
}
\description{
Load tidy data sources configuration
}
\keyword{internal}
