% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhstt-package.R
\docType{package}
\name{nhstt-package}
\alias{nhstt}
\alias{nhstt-package}
\title{nhstt: NHS Talking Therapies for anxiety and depression reports in R}
\description{
Analysis-ready datasets of publicly available reports from NHS Talking Therapies for anxiety and depression (NHS TT).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bennettoxford/nhstt}
  \item \url{https://bennettoxford.github.io/nhstt/}
  \item Report bugs at \url{https://github.com/bennettoxford/nhstt/issues}
}

}
\author{
\strong{Maintainer}: Milan Wiedemann \email{milan.wiedemann@gmail.com} (\href{https://orcid.org/0000-0003-1991-282X}{ORCID})

Authors:
\itemize{
  \item Milan Wiedemann \email{milan.wiedemann@gmail.com} (\href{https://orcid.org/0000-0003-1991-282X}{ORCID})
  \item Lola Ojedele \email{lolaojedele@gmail.com} (\href{https://orcid.org/0009-0008-4895-6913}{ORCID})
  \item Viveck Kingsley \email{viveck.kingsley@phc.ox.ac.uk} (\href{https://orcid.org/0009-0003-0864-7132}{ORCID})
}

Other contributors:
\itemize{
  \item Wellcome Trust [funder]
}

}
\keyword{internal}
