% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-utils.R
\name{parse_annual_period_bounds}
\alias{parse_annual_period_bounds}
\title{Parse annual period bounds}
\usage{
parse_annual_period_bounds(period)
}
\arguments{
\item{period}{Character, specifying financial year (e.g., "2023-24")}
}
\value{
List with start and end dates
}
\description{
Converts financial year code to start/end dates
}
\keyword{internal}
