% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-utils.R
\name{pivot_longer_measures}
\alias{pivot_longer_measures}
\title{Pivot measures to long format}
\usage{
pivot_longer_measures(data_list, pivot_config)
}
\arguments{
\item{data_list}{Named list, specifying tibbles (e.g., list("2023-24" = df1, "2024-25" = df2))}

\item{pivot_config}{List, specifying pivot configuration with elements:
\itemize{
\item id_cols: Character vector, ID columns to preserve
\item measure_cols: Character vector, measure columns to pivot
\item sep: Character, regex pattern to separate measure names
\item names_to: Character vector, output column names for separated parts
}}
}
\value{
Tibble in long format
}
\description{
Pivot measures to long format
}
\keyword{internal}
