% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{read_raw_downloads_json}
\alias{read_raw_downloads_json}
\title{Read raw downloads metadata}
\usage{
read_raw_downloads_json(frequency)
}
\arguments{
\item{frequency}{Character, specifying report frequency ("annual" or "monthly")}
}
\value{
List with download metadata, or empty list if file doesn't exist
}
\description{
Read raw downloads metadata
}
\keyword{internal}
