% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-utils.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename columns}
\usage{
rename_columns(df, rename_mapping)
}
\arguments{
\item{df}{Tibble, specifying data with columns to rename}

\item{rename_mapping}{Named character vector, specifying old_name = new_name mapping}
}
\value{
Tibble with renamed columns
}
\description{
Renames columns using a mapping (for year-specific inconsistencies)
}
\keyword{internal}
