% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_validate.R
\name{validate_dataset}
\alias{validate_dataset}
\title{Validate dataset}
\usage{
validate_dataset(df, dataset_name = "dataset")
}
\arguments{
\item{df}{Data frame to validate}

\item{dataset_name}{Name of dataset for error messages}
}
\value{
Invisible TRUE if valid
}
\description{
Validate dataset
}
\keyword{internal}
