% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-config.R
\name{validate_dataset}
\alias{validate_dataset}
\title{Validate dataset name}
\usage{
validate_dataset(dataset, frequency)
}
\arguments{
\item{dataset}{Character, specifying dataset name}

\item{frequency}{Character, specifying report frequency ("annual" or "monthly")}
}
\value{
Invisible TRUE if valid, aborts otherwise
}
\description{
Validate dataset name
}
\keyword{internal}
