% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-config.R
\name{validate_dataset}
\alias{validate_dataset}
\title{Validate dataset name}
\usage{
validate_dataset(dataset)
}
\arguments{
\item{dataset}{Character, specifying dataset name to validate (e.g., "key_measures", "activity_performance")}
}
\value{
Invisible TRUE if valid, aborts otherwise
}
\description{
Validate dataset name
}
\keyword{internal}
